+++ src/pages/mypage/components/BottomSection/bottomSection.css.ts
