+++ src/pages/mypage/components/mypageReservationDetail/apis/queries.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
-import { getReservationsDetail } from '@/pages/mypage/mypageReservationDetail/apis/axios';
-import type { ReservationDetailType } from '@/pages/mypage/mypageReservationDetail/types/api';
+import { getReservationsDetail } from '@/pages/mypage/components/mypageReservationDetail/apis/axios';
+import type { ReservationDetailType } from '@/pages/mypage/components/mypageReservationDetail/types/api';
 import { QUERY_KEYS } from '@/shared/constants/queryKey';
 
 export const useGetReservationsDetail = (lessonId: number) => {
