+++ src/pages/mypage/components/mypageReservationDetail/mypageReservationDetail.css.ts
