+++ src/pages/mypage/components/mypageReservationDetail/types/api.ts
