+++ src/routes/router.tsx
@@ -10,8 +10,8 @@ import ClassRegister from '@/pages/instructor/classRegister/ClassRegister';
 import ClassRegisterCompletion from '@/pages/instructor/classRegisterCompletion/ClassRegisterCompletion';
 import InstructorRegister from '@/pages/instructorRegister/InstructorRegister';
 import Login from '@/pages/login/Login';
-import MyPageReservation from '@/pages/mypage/mypageReservation/MypageReservation';
-import MyPageReservationDetail from '@/pages/mypage/mypageReservationDetail/MypageReservationDetail';
+import MyPageReservation from '@/pages/mypage/components/mypageReservation/MypageReservation';
+import MyPageReservationDetail from '@/pages/mypage/components/mypageReservationDetail/MypageReservationDetail';
 import Onboarding from '@/pages/onboarding/Onboarding';
 import Reservation from '@/pages/reservation/Reservation';
 import { CheckoutPage } from '@/pages/reservation/components/TossPayments/CheckOut/CheckOut';
